Delete merged away segment files in periodic flow

+++ server/src/main/java/org/opensearch/index/shard/RemoteStoreRefreshListener.java
@@ -11,15 +11,22 @@ package org.opensearch.index.shard;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.apache.lucene.index.SegmentInfos;
 import org.apache.lucene.search.ReferenceManager;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
+import org.apache.lucene.store.IndexOutput;
+import org.apache.lucene.util.Version;
 
 import java.io.IOException;
 import java.nio.file.NoSuchFileException;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 
 /**
@@ -27,17 +34,20 @@ import java.util.stream.Collectors;
  */
 public class RemoteStoreRefreshListener implements ReferenceManager.RefreshListener {
 
+    private static final String INTERMEDIATE_SEGMENTINFOS_FILENAME = "intermediate_segments_N";
     private final Directory storeDirectory;
     private final Directory remoteDirectory;
     // ToDo: This can be a map with metadata of the uploaded file as value of the map (GitHub #3398)
     private final Set<String> filesUploadedToRemoteStore;
     private static final Logger logger = LogManager.getLogger(RemoteStoreRefreshListener.class);
+    private final ScheduledThreadPoolExecutor executor;
 
     public RemoteStoreRefreshListener(Directory storeDirectory, Directory remoteDirectory) throws IOException {
         this.storeDirectory = storeDirectory;
         this.remoteDirectory = remoteDirectory;
         // ToDo: Handle failures in reading list of files (GitHub #3397)
         this.filesUploadedToRemoteStore = new HashSet<>(Arrays.asList(remoteDirectory.listAll()));
+        this.executor = new ScheduledThreadPoolExecutor(1);
     }
 
     public Directory getStoreDirectory() {
@@ -56,12 +66,14 @@ public class RemoteStoreRefreshListener implements ReferenceManager.RefreshListe
     /**
      * Upload new segment files created as part of the last refresh to the remote segment store.
      * The method also deletes segment files from remote store which are not part of local filesystem.
+     *
      * @param didRefresh true if the refresh opened a new reference
      * @throws IOException in case of I/O error in reading list of local files
      */
     @Override
     public void afterRefresh(boolean didRefresh) throws IOException {
         if (didRefresh) {
+            AtomicBoolean uploadSuccess = new AtomicBoolean(true);
             Set<String> localFiles = Arrays.stream(storeDirectory.listAll()).collect(Collectors.toSet());
             localFiles.stream().filter(file -> !filesUploadedToRemoteStore.contains(file)).forEach(file -> {
                 try {
@@ -73,24 +85,36 @@ public class RemoteStoreRefreshListener implements ReferenceManager.RefreshListe
                         e
                     );
                 } catch (IOException e) {
+                    uploadSuccess.set(false);
                     // ToDO: Handle transient and permanent un-availability of the remote store (GitHub #3397)
                     logger.warn(() -> new ParameterizedMessage("Exception while uploading file {} to the remote segment store", file), e);
                 }
             });
 
-            Set<String> remoteFilesToBeDeleted = new HashSet<>();
-            // ToDo: Instead of deleting files in sync, mark them and delete in async/periodic flow (GitHub #3142)
-            filesUploadedToRemoteStore.stream().filter(file -> !localFiles.contains(file)).forEach(file -> {
-                try {
-                    remoteDirectory.deleteFile(file);
-                    remoteFilesToBeDeleted.add(file);
-                } catch (IOException e) {
-                    // ToDO: Handle transient and permanent un-availability of the remote store (GitHub #3397)
-                    logger.warn(() -> new ParameterizedMessage("Exception while deleting file {} from the remote segment store", file), e);
-                }
-            });
+            if (uploadSuccess.get()) {
+                IndexOutput indexOutput = storeDirectory.createOutput(INTERMEDIATE_SEGMENTINFOS_FILENAME, IOContext.DEFAULT);
+                SegmentInfos segmentInfos = new SegmentInfos(Version.LATEST.major);
+                segmentInfos.write(indexOutput);
+                indexOutput.close();
+                storeDirectory.sync(Collections.singleton(INTERMEDIATE_SEGMENTINFOS_FILENAME));
+                remoteDirectory.copyFrom(storeDirectory, INTERMEDIATE_SEGMENTINFOS_FILENAME, INTERMEDIATE_SEGMENTINFOS_FILENAME, IOContext.DEFAULT);
 
-            remoteFilesToBeDeleted.forEach(filesUploadedToRemoteStore::remove);
+                executor.schedule(() -> {
+                    Set<String> remoteFilesToBeDeleted = new HashSet<>();
+                    // ToDo: Instead of deleting files in sync, mark them and delete in async/periodic flow (GitHub #3142)
+                    filesUploadedToRemoteStore.stream().filter(file -> !localFiles.contains(file)).forEach(file -> {
+                        try {
+                            remoteDirectory.deleteFile(file);
+                            remoteFilesToBeDeleted.add(file);
+                        } catch (IOException e) {
+                            // ToDO: Handle transient and permanent un-availability of the remote store (GitHub #3397)
+                            logger.warn(() -> new ParameterizedMessage("Exception while deleting file {} from the remote segment store", file), e);
+                        }
+                    });
+                    remoteFilesToBeDeleted.forEach(filesUploadedToRemoteStore::remove);
+                }, 15, TimeUnit.MINUTES);
+            }
         }
     }
 }
+
